Update LAPSLookup.ps1
Before when we looked up the password expiry date, if it was blank in AD, the script would error before printing a password as we tried to parse the datetime in place as part of the return value. Now we check for blank values and only parse it if its not blank.